AP-2718: Refactor the CCMSRestartSubmissions namespace
