+++ lib/tasks/restart_ccms_submissions.rake
@@ -1,6 +1,6 @@
 namespace :ccms do
   desc 'Restart CCMS submissions after CCMS have been turned off'
   task restart_submissions: :environment do
-    CCMSRestartSubmissions.call
+    CCMS::RestartSubmissions.call
   end
 end
